Created root object
